manga-application-server/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.32"
sqlalchemy-utils = "^0.41.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aiosqlite = "^0.20.0"
pillow = "^10.4.0"
